python/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.38.0
h5py==3.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keras-nightly==2.6.0.dev2021052200
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pywin32==300
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
tb-nightly==2.6.0a20210521
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032601
tf-nightly==2.6.0.dev20210521
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
